Updated `deployTOFT`
- Using netwrok name instead of chainid
- TOFT constructor param type update
- new util

+++ contracts/TapiocaOFT.sol
@@ -27,7 +27,7 @@ contract TapiocaOFT is OFT {
         string memory _name,
         string memory _symbol,
         uint8 _decimal,
-        uint8 _mainChainID
+        uint16 _mainChainID
     )
         OFT(
             string(abi.encodePacked('TapiocaWrapper-', _name)),
